Extract options
